+++ whatsapp_mail_messaging/__manifest__.py
@@ -21,7 +21,7 @@
 #############################################################################
 {
     'name': 'Odoo Whatsapp Connector',
-    'version': '13.0.1.0.0',
+    'version': '13.0.1.1.0',
     'category': 'Extra Tools',
     'summary': """Odoo Whatsapp Connector For Sales, Invoice, and Floating button in Website""",
     'description': """Added options for sending Whatsapp messages and Mails in systray bar,sale order, invoices,
@@ -38,6 +38,8 @@
         'views/sale_order_inherited.xml',
         'views/account_move_inherited.xml',
         'views/website_inherited.xml',
+        'views/res_company_views.xml',
+        'views/res_config_settings_views.xml',
         'wizard/wh_message_wizard.xml',
         'wizard/portal_share_inherited.xml',
         'security/ir.model.access.csv',