+++ whatsapp_mail_messaging/model/account_move.py
@@ -29,9 +29,11 @@ class Account(models.Model):
     def action_send_whatsapp(self):
         compose_form_id = self.env.ref('whatsapp_mail_messaging.whatsapp_message_wizard_form').id
         ctx = dict(self.env.context)
-        message = "Hi" + " " + self.partner_id.name + ',' + '\n' + "Here is your invoice" + ' ' + self.name + ' ' + "amounting" + ' ' + str(
+        message_template = self.company_id.whatsapp_message
+        default_message = "Hi" + " " + self.partner_id.name + ',' + '\n' + "Here is your invoice" + ' ' + self.name + ' ' + "amounting" + ' ' + str(
             self.amount_total) + self.currency_id.symbol + ' ' + "from " + self.company_id.name + ". Please remit payment at your earliest convenience. " + '\n' + \
                   "Please use the following communication for your payment" + ' ' + self.name
+        message = message_template if message_template else default_message
         ctx.update({
             'default_message': message,
             'default_partner_id': self.partner_id.id,