+++ whatsapp_mail_messaging/model/sale_order.py
@@ -29,8 +29,10 @@ class Sale(models.Model):
     def action_send_whatsapp(self):
         compose_form_id = self.env.ref('whatsapp_mail_messaging.whatsapp_message_wizard_form').id
         ctx = dict(self.env.context)
-        message = "Hi" + " " + self.partner_id.name + ',' + '\n' + "Your quotation" + ' ' + self.name + ' ' + "amounting" + ' ' + str(
+        message_template = self.company_id.whatsapp_message
+        default_message = "Hi" + " " + self.partner_id.name + ',' + '\n' + "Your quotation" + ' ' + self.name + ' ' + "amounting" + ' ' + str(
             self.amount_total) + self.currency_id.symbol + ' ' + "is ready for review.Do not hesitate to contact us if you have any questions."
+        message = message_template if message_template else default_message
         ctx.update({
             'default_message': message,
             'default_partner_id': self.partner_id.id,