+++ whatsapp_mail_messaging/doc/changelog.md
@@ -3,8 +3,11 @@
 #### 22.06.2021
 #### Version 13.0.1.0.0
 #### ADD
-- Initial commit
+- Initial commit Odoo Whatsapp Connector
 
-
+#### 20.05.2024
+#### Version 13.0.1.1.0
+#### UPDT
+- Added whatsapp message template feature.